wxNviz: zexag value must be saved as float
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@49377 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ gui/wxpython/core/workspace.py
@@ -484,7 +484,7 @@ class ProcessWorkspaceFile:
         node_zexag = node_view.find('z-exag')
         view['z-exag'] = {}
         iview['z-exag'] = {}
-        view['z-exag']['value'] = self.__processLayerNvizNode(node_zexag, 'value', int)
+        view['z-exag']['value'] = self.__processLayerNvizNode(node_zexag, 'value', float)
         view['z-exag']['min'] = self.__processLayerNvizNode(node_zexag, 'min', int)
         view['z-exag']['max'] = self.__processLayerNvizNode(node_zexag, 'max', int)
         iview['z-exag']['original'] = self.__processLayerNvizNode(node_zexag, 'original', float)
@@ -919,10 +919,10 @@ class WriteWorkspaceFile(object):
         # z-exag
         self.file.write('%s<z-exag>\n' % (' ' * self.indent))
         self.indent += 4
-        self.file.write('%s<value>%d</value>\n' % (' ' * self.indent, view['z-exag']['value']))
+        self.file.write('%s<value>%.2f</value>\n' % (' ' * self.indent, view['z-exag']['value']))
         self.file.write('%s<min>%d</min>\n' % (' ' * self.indent, view['z-exag']['min']))
         self.file.write('%s<max>%d</max>\n' % (' ' * self.indent, view['z-exag']['max']))
-        self.file.write('%s<original>%d</original>\n' % (' ' * self.indent, iview['z-exag']['original']))
+        self.file.write('%s<original>%.2f</original>\n' % (' ' * self.indent, iview['z-exag']['original']))
         self.indent -= 4
         self.file.write('%s</z-exag>\n' % (' ' * self.indent))
         # focus (look here)
